@@ -44,33 +44,33 @@ public interface ITimeEntries
 		CancellationToken cancellationToken
 		);
 
-    /// <summary>
-    /// Load time entry by ID that is accessible by the current user.
-    /// https://engineering.toggl.com/docs/api/time_entries#get-get-a-time-entry-by-id
-    /// </summary>
-    /// <param name="timeEntryId">TimeEntry ID</param>
-    /// <param name="cancellationToken">The cancellation token</param>
-    /// <returns></returns>
-    [Get("/api/v9/me/time_entries/{time_entry_id}")]
+	/// <summary>
+	/// Load time entry by ID that is accessible by the current user.
+	/// https://engineering.toggl.com/docs/api/time_entries#get-get-a-time-entry-by-id
+	/// </summary>
+	/// <param name="timeEntryId">TimeEntry ID</param>
+	/// <param name="cancellationToken">The cancellation token</param>
+	/// <returns></returns>
+	[Get("/api/v9/me/time_entries/{time_entry_id}")]
 	Task<TimeEntry> GetAsync(
 		[AliasAs("time_entry_id")] long timeEntryId,
 		CancellationToken cancellationToken
 		);
 
-    /// <summary>
-    /// In short: http://tools.ietf.org/html/rfc6902 and http://tools.ietf.org/html/rfc6901 with some additions. Patch will be executed partially when there are errors with some records. No transaction, no rollback.
-    /// https://engineering.toggl.com/docs/api/time_entries#patch-bulk-editing-time-entries
-    /// </summary>
-    /// <param name="workspaceId">Numeric ID of the workspace</param>
-    /// <param name="timeEntryIds">Numeric IDs of time_entries, separated by comma. E.g.: 204301830,202700150,202687559. The limit is 100 IDs per request.</param>
-    /// <param name="array">Array of batch operations</param>
-    /// <param name="cancellationToken">The cancellation token</param>
-    /// <returns></returns>
-    [Patch("/api/v9/workspaces/{workspace_id}/time_entries/{time_entry_ids}")]
-    Task PatchAsync(
-        [AliasAs("workspace_id")] long workspaceId,
-        [AliasAs("time_entry_ids")] string timeEntryIds,
-        [Body] ICollection<TimeEntryBulkEditDto> array,
-        CancellationToken cancellationToken
-        );
+	/// <summary>
+	/// In short: http://tools.ietf.org/html/rfc6902 and http://tools.ietf.org/html/rfc6901 with some additions. Patch will be executed partially when there are errors with some records. No transaction, no rollback.
+	/// https://engineering.toggl.com/docs/api/time_entries#patch-bulk-editing-time-entries
+	/// </summary>
+	/// <param name="workspaceId">Numeric ID of the workspace</param>
+	/// <param name="timeEntryIds">Numeric IDs of time_entries, separated by comma. E.g.: 204301830,202700150,202687559. The limit is 100 IDs per request.</param>
+	/// <param name="array">Array of batch operations</param>
+	/// <param name="cancellationToken">The cancellation token</param>
+	/// <returns></returns>
+	[Patch("/api/v9/workspaces/{workspace_id}/time_entries/{time_entry_ids}")]
+	Task PatchAsync(
+		[AliasAs("workspace_id")] long workspaceId,
+		[AliasAs("time_entry_ids")] string timeEntryIds,
+		[Body] ICollection<TimeEntryBulkEditDto> array,
+		CancellationToken cancellationToken
+		);
 }
