@@ -10,9 +10,9 @@ public class TimeEntryBulkEditDto
 	[JsonPropertyName("op")]
 	public required string op { get; set; }
 
-    [JsonPropertyName("path")]
-    public required string path { get; set; }
+	[JsonPropertyName("path")]
+	public required string path { get; set; }
 
-    [JsonPropertyName("value")]
-    public required object value { get; set; }
+	[JsonPropertyName("value")]
+	public required object value { get; set; }
 }
